rename vizwit-gist to vizwit-loader, add local file fetching, update readme
