+++ README.md
@@ -9,7 +9,7 @@ An interactive data visualization tool. VizWit uses a [JSON config file](https:/
 interactive charts that cross-filter one another. It currently supports data hosted in a Socrata Open Data portal, which includes 
 cities such as [Philadelphia](http://opendataphilly.org), [Chicago](https://data.cityofchicago.org/), [San Francisco](https://data.sfgov.org/)
 and many others. However, interactions with Socrata have been [abstracted](src/scripts/collections/socrata.js) to allow
-for other data providers to be written (in theory).
+for [other data providers](https://github.com/timwis/vizwit/wiki/Adding-a-provider) to be written.
 
 [![screencast](http://i.imgur.com/4gTXNFK.gif)](http://vizw.it/?gist=51db593dc0537d1a3f05)
 
@@ -60,16 +60,17 @@ layout is generated by Gridstack. The application is compiled using Browserify. 
 # How it works
 [vizwit.js](src/scripts/vizwit.js) is the primary chart-generating module. It takes a `container` selector and a `config` object and
 uses them to initialize a view (bar chart, map, table, etc.). The views are passed a `collection` and a `filteredCollection`, which are
-identical and allow the view to query the data provider (Socrata). Views generate the chart/map/table/etc. and listen for user interactions
+identical and allow the view to query the data provider. Views generate the chart/map/table/etc. and listen for user interactions
 on them. On an interaction, the view triggers the global event object with the filter (ie. `state=PA`), and all views that use the same 
 dataset receive the event and use their `filteredCollection` to query the data provider with the filter passed through the event. By
 using a separate collection, VizWit can show the "filtered value" and the "original value" side-by-side.
 
-The actual entry point is either [layout.js](src/scripts/layout.js) or [embed.js](src/scripts/embed.js). layout.js fetches a gist, reads
-its configuration, and creates a layout on the page. Then for each chart in the configuration, it calls `init()` from vizwit.js as 
-described above. embed.js simply finds the parent container of the `<script>` tag that was embedded, and calls `init()` from vizwit.js
-inside of it. This way of embedding allows the VizWit library to only be loaded once on the page with as many charts embedded in any
-element ([example](http://vizw.it/embed-demo.html)).
+The actual entry point is either [vizwit-loader.js](src/scripts/vizwit-loader.js) or [vizwit-embed.js](src/scripts/vizwit-embed.js). 
+vizwit-loader.js fetches a gist or local file, reads its configuration, and passes it to [layout.js](src/scripts/layout.js] to create a 
+layout on the page. Then for each chart in the configuration, layout.js calls `init()` from vizwit.js as described above. vizwit-embed.js 
+simply finds the parent container of the `<script>` tag that was embedded, and calls `init()` from vizwit.js inside of it. This way of 
+embedding allows the VizWit library to only be loaded once on the page with as many charts embedded in any element 
+([example](http://vizw.it/embed-demo.html)).
 
 # Development
 * After cloning the repo, use `npm install` to install dependencies
