+++ src/index.html
@@ -42,6 +42,6 @@
 			<div class="grid-stack grid-stack-12" id="page-content"></div>
 		</div>
 		
-		<script src="scripts/vizwit-gist.js"></script>
+		<script src="scripts/vizwit-loader.js"></script>
 	</body>
 </html>
